Handles errors correctly
It now sends back {action: 'error', type: <action>} -- where <action> is
the action that was requested, in uppercase -- when an error happens in
the actions. It allows the extension to see when someone isn't trusted
if it wanted to encrypt something, for example.

Also don't crash when the communication is off, just accept that you're
not the center of the world, and sometimes chromium doesn't want to
speak with you anymore.

More seriously, it stops the errors when refreshing the extension.